# api/docs/swagger.yaml
definitions:
...
  genproto.ContextRes:
    properties:
      historical:
        $ref: '#/definitions/genproto.HistoricalEventsGetAllRes'
      personal:
        $ref: '#/definitions/genproto.PersonalEventsGetAllRes'
    type: object
...
  genproto.HistoricalEventsCreateReq:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  genproto.HistoricalEventsGetAllRes:
    properties:
      count:
        type: integer
      events:
        items:
          $ref: '#/definitions/genproto.HistoricalEventsRes'
        type: array
    type: object
  genproto.HistoricalEventsGetByIdRes:
    properties:
      event:
        $ref: '#/definitions/genproto.HistoricalEventsRes'
    type: object
  genproto.HistoricalEventsRes:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  genproto.MediasCreateReq:
    properties:
      memory_id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  genproto.MediasGetAllRes:
    properties:
      count:
        type: integer
      medias:
        items:
          $ref: '#/definitions/genproto.MediasRes'
        type: array
    type: object
  genproto.MediasGetByIdRes:
    properties:
      media:
        $ref: '#/definitions/genproto.MediasRes'
    type: object
  genproto.MediasRes:
    properties:
      id:
        type: string
      memory_id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
  genproto.PersonalEventsGetAllRes:
    properties:
      count:
        type: integer
      events:
        items:
          $ref: '#/definitions/genproto.PersonalEventsRes'
        type: array
    type: object
...
  genproto.PersonalEventsRes:
    properties:
      date:
        type: string
      id:
        type: string
      preview:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /memory/{id}/media:
    post:
      consumes:
      - application/json
      description: Create a new media
      parameters:
      - description: Media data
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/genproto.MediasCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Media created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create media
      tags:
      - media
  /memory/{id}/media/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a media by ID
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Media deleted
          schema:
            type: string
        "400":
          description: Invalid media ID
          schema:
            type: string
        "404":
          description: Media not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete media
      tags:
      - media
    get:
      consumes:
      - application/json
      description: Get a media
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.MediasGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get media
      tags:
      - media
    put:
      consumes:
      - application/json
      description: Update an existing media
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      - description: Type
        in: query
        name: type
        type: string
      - description: Url
        in: query
        name: url
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Media updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Media not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update media
      tags:
      - media
  /memory/{id}/media/all:
    get:
      consumes:
      - application/json
      description: Get all medias
      parameters:
      - description: MemoryId
        in: query
        name: memory_id
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.MediasGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all medias
      tags:
      - media
...
  /timeline/context/{date}:
    get:
      consumes:
      - application/json
      description: Get all event
      parameters:
      - description: Date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.ContextRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all event
      tags:
      - context
...
  /timeline/historical:
    post:
      consumes:
      - application/json
      description: Create a new event
      parameters:
      - description: Event data
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/genproto.HistoricalEventsCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Event created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create event
      tags:
      - historical
  /timeline/historical/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a event by ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event deleted
          schema:
            type: string
        "400":
          description: Invalid event ID
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete event
      tags:
      - historical
    get:
      consumes:
      - application/json
      description: Get a event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.HistoricalEventsGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get event
      tags:
      - historical
    put:
      consumes:
      - application/json
      description: Update an existing event
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      - description: Title
        in: query
        name: title
        type: string
      - description: Description
        in: query
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Event not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update event
      tags:
      - historical
  /timeline/historical/all:
    get:
      consumes:
      - application/json
      description: Get all event
      parameters:
      - description: Date
        in: query
        name: date
        type: string
      - description: Category
        in: query
        name: category
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.HistoricalEventsGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all event
      tags:
      - historical
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"